The previous version was better
As it does not contradict existing research. (Even though I think this
is the wrong approach...)

Anyway, it's debatable, so I will keep the same approach of the
previous research.
To solve this problem unambiguously I need a tremendously big dataset,
and I'm not willing to do so now, given that the contradiction has
very minimal consequences.