Greatly sped up data extraction
Instead of iterating giant lists of opcodes through python, the new code
delegates r2 (which is compiled and runs faster. A LOT faster, from
several minutes to several seconds for very big files).

If the amount of files is big, the extraction can still take a lot of
times, for this reason a progress bar have been added.