Speedup in the preprocess phase
More than 60x faster in the current dataset, just by using proper buffering.
Shuffling and Writing are still the bottleneck however, but there's not
much I can do in that direction without loading everything in memory
(and that's exactly why I rewrote the dataset, to avoid having
everything in memory).
So, to avoid useless shuffling/deduplicating/splitting, I added a new
flag called --incomplete, to signal when the dataset is being built as
partial, so these slow procedures won't run (because they are needed
only once, at the end).